requirements.txt
flask
flask-restful
requests
radon
gitpython
werkzeug
lizard